Make prompt generator based on database type - for multi-runs
